+++ LegitFlix.Plugin/obj/Release/net9.0/LegitFlix.Plugin.AssemblyInfo.cs
@@ -14,7 +14,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Release")]
 [assembly: System.Reflection.AssemblyDescriptionAttribute("Total UI replacement for Jellyfin.")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.31")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+3bf6078119df919d758680a3a0ce1013ba39deaf")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+83ed1cb04e3cc82ed45c348fe45d03b6b4a5ffa5")]
 [assembly: System.Reflection.AssemblyProductAttribute("LegitFlix.Plugin")]
 [assembly: System.Reflection.AssemblyTitleAttribute("LegitFlix.Plugin")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.31")]